mcmugane_Web/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pandas==2.2.2
numpy==1.26.4
pydantic==2.7.1
alpaca-py==0.42.0
python-dotenv==1.0.1
tabulate==0.9.0
orjson==3.10.3
